Refactor UI file structure and merge RedactedText
Moved UI source files into feature-specific subdirectories for improved organization. Merged the RedactedText view into Controls/RedactedText.swift and removed the duplicate file.
