+++ Sources/Scout/UI/Activity/ActivityRow.swift
