+++ Sources/Scout/UI/Analytics/AnalyticsView.Suggestion.swift
