+++ Sources/Scout/UI/Controls/Header.swift
