+++ Sources/Scout/UI/Controls/Message.swift
