+++ Sources/Scout/UI/Database/DatabaseController+Sample.swift
