+++ Sources/Scout/UI/Event/EventProvider.swift
