+++ Sources/Scout/UI/Event/EventView.Sections.swift
