+++ Sources/Scout/UI/Event/Range/RangeControl.swift
