+++ Sources/Scout/UI/Event/Range/RangeLabel.swift
