+++ Sources/Scout/UI/Event/Stat/Period.swift
