+++ Sources/Scout/UI/Event/Stat/PeriodPicker.swift
