+++ Sources/Scout/UI/Event/Stat/StatModel+Move.swift
