+++ Sources/Scout/UI/Event/Stat/StatModel.swift
