+++ Sources/Scout/UI/Event/Stat/StatRow.swift
