+++ Sources/Scout/UI/Event/Stat/StatView.swift
