+++ Sources/Scout/UI/Filter/Criteria.swift
