+++ Sources/Scout/UI/History/HistoryFilter.swift
