+++ Sources/Scout/UI/History/HistoryView.swift
