+++ Sources/Scout/UI/Home/HomeView.Sections.swift
