+++ Sources/Scout/UI/Utilities/Box.swift
