+++ Sources/Scout/UI/Utilities/Calendar+QueryRange.swift
